+++ crates/vue_oxlint_parser/Cargo.toml
@@ -22,6 +22,7 @@ oxc_estree_tokens = { workspace = true }
 oxc_ast_visit = { workspace = true }
 
 memchr = { workspace = true }
+regex = { workspace = true }
 rustc-hash = { workspace = true }
 
 [dev-dependencies]
